Update README.md
For my project, I wanted to understand what makes a song popular. To achieve this, I analyzed a dataset of over 18,000 songs from 2000-2020 and looked at various song attributes such as valence, danceability, and key to predict song popularity.

To make sense of this data, I used a variety of data visualization techniques such as histograms, scatterplots, line graphs, and box plots. For example, I used histograms to visualize the shape of the data and scatterplots to show how song duration and popularity were related.

I also used line graphs to demonstrate how the energy level and audio valence of a song impacted its popularity, and I used box plots to study how the audio mode and key of a song related to its danceability. Additionally, I used regression to predict song popularity based on a combination of danceability, liveness, audio valence, and energy.

By examining the data in this way, I was able to identify patterns and trends that helped me better understand the factors that contribute to the popularity of a song. Overall, my project provides valuable insights into the music industry and could be useful for musicians, producers, and anyone else interested in the factors that make a song popular.